Heap, Heap Min and heap Max
Update: heap and Heap max
Creation: Heap Min
