+++ Data/Heap.java
@@ -2,35 +2,31 @@ package Data;
 
 public class Heap {
 
-    DynamicList HeapArray = null;
+    public DynamicList array = null;
 
     public Heap() {
-        this.HeapArray = new DynamicList();
+        this.array = new DynamicList();
     }
 
     public int getSize() {
-        return this.HeapArray.index;
+        return this.array.index;
     }
 
     public int getMaxSize() {
-        return this.HeapArray.size;
-    }
-
-    public int getMax() {
-        return this.HeapArray.list[0];
+        return this.array.size;
     }
 
     public int parent(int index) throws Exception {
-        int indexParent = ((index+1) / 2)-1;
+        int indexParent = ((index + 1) / 2)-1;
         if (indexParent >= this.getSize()){
-            throw new Exception("No parent: Index "+indexParent+" out of bounds for size "+this.getSize());
+            throw new Exception("No parent: Index "+ indexParent +" out of bounds for size "+this.getSize());
         } else {
             return indexParent;
         }
     }
 
     public int leftChild(int index) throws Exception {
-        int indexChildL = ((index+1) * 2)-1;
+        int indexChildL = ((index + 1) * 2)-1;
         if (indexChildL >= this.getSize()){
             throw new Exception("No left child: Index "+indexChildL+" out of bounds for size "+this.getSize());
         } else {
@@ -39,7 +35,7 @@ public class Heap {
     }
 
     public int rightChild(int index) throws Exception {
-        int indexChildR = (((index+1) * 2) + 1)-1;
+        int indexChildR = (((index + 1) * 2) + 1)-1;
         if (indexChildR >= this.getSize()){
             throw new Exception("No right child: Index "+indexChildR+" out of bounds for size "+this.getSize());
         } else {
@@ -48,24 +44,12 @@ public class Heap {
     }
 
     public int levelNode(int index) {
-        int i = index+1;
+        int i = index + 1;
         return (int) ((Math.log(i) / Math.log(2)) + 1);
     }
 
-    public void swiftUp(int index) throws Exception {
-        if (index > 0){
-            if (this.HeapArray.list[this.parent(index)] < this.HeapArray.list[index]){
-                int temp = this.HeapArray.list[this.parent(index)];
-                this.HeapArray.list[this.parent(index)] = this.HeapArray.list[index];
-                this.HeapArray.list[index] = temp;
-                this.swiftUp(this.parent(index));
-            }
-        }
-    }
-
     public void insert(int key) throws Exception {
-        this.HeapArray.pushBack(key);
-        this.swiftUp(this.getSize()-1);
+        this.array.pushBack(key);
     }
 
     public void lineToInsert(String data) throws Exception {
@@ -76,14 +60,14 @@ public class Heap {
     }
 
     public String levelOrder() throws Exception {
-        if (this.HeapArray.empty()){
+        if (this.array.empty()){
             throw new Exception("Empty Heap");
         } else {
             int l = 1;
             String result = "";
             for (int i = 0; i<this.getSize(); i++){
-                int cur_element = this.HeapArray.list[i];
-                int cur_level = this.levelNode(i);
+                int cur_element = this.array.list[i];
+                int cur_level = this.levelNode(cur_element);
                 if ((l+1) == cur_level){
                     result += "\n"+cur_element + " ";
                     l += 1;
@@ -95,99 +79,86 @@ public class Heap {
         }
     }
 
-    public void swiftDown(int index) throws Exception {
-        int maxIndex = index;
-        try{
-            int l = this.leftChild(index);
-            if (l < this.getSize() && this.HeapArray.list[l] > this.HeapArray.list[maxIndex]){
-                maxIndex = l;
+    public String inOrder(int index) throws Exception {
+        if (index == this.getSize()){
+            return "";
+        } else {
+            String result = "";
+            try{
+                result += this.inOrder(this.leftChild(index));
+            } catch (Exception ignored){}
+            int num = this.array.list[index];
+            if (num != -1){
+                result += num + " ";
             }
-        } catch (Exception ignored){}
-        try{
-            int r = this.rightChild(index);
-            if (r < this.getSize() && this.HeapArray.list[r] > this.HeapArray.list[maxIndex]){
-                maxIndex = r;
-            }
-        } catch (Exception ignored){}
-        if (index != maxIndex){
-            int temp = this.HeapArray.list[index];
-            this.HeapArray.list[index] = this.HeapArray.list[maxIndex];
-            this.HeapArray.list[maxIndex] = temp;
-            this.swiftDown(maxIndex);
+            try{
+                result += this.inOrder(this.rightChild(index));
+            } catch (Exception ignored){}
+            return result;
         }
     }
 
-    public int extractMax() throws Exception {
-        int result = this.HeapArray.list[0];
-        int last = this.HeapArray.list[this.getSize()-1];
-        this.HeapArray.list[0] = last;
-        this.HeapArray.index -= 1;
-        this.swiftDown(0);
-        return result;
+    public String inOrder() throws Exception {
+        return this.inOrder(0);
+    }
+
+    public String preOrder(int index) throws Exception {
+        if (index == this.getSize()){
+            return "";
+        } else {
+            String result = "";
+            int num = this.array.list[index];
+            if (num != -1){
+                result += num + " ";
+            }
+            try{
+                result += this.preOrder(this.leftChild(index));
+            } catch (Exception ignored){}
+            try{
+                result += this.preOrder(this.rightChild(index));
+            } catch (Exception ignored){}
+            return result;
+        }
+    }
+
+    public String preOrder() throws Exception {
+        return this.preOrder(0);
+    }
+
+    public String posOrder(int index) throws Exception {
+        if (index == this.getSize()){
+            return "";
+        } else {
+            String result = "";
+            try{
+                result += this.posOrder(this.leftChild(index));
+            } catch (Exception ignored){}
+            try{
+                result += this.posOrder(this.rightChild(index));
+            } catch (Exception ignored){}
+            int num = this.array.list[index];
+            if (num != -1){
+                result += num + " ";
+            }
+            return result;
+        }
+    }
+
+    public String posOrder() throws Exception {
+        return this.posOrder(0);
     }
 
     public void remove(int index) throws Exception {
-        this.HeapArray.list[index] = this.getMax()+1;
-        this.swiftUp(index);
-        this.extractMax();
+        if (index == (this.getSize()-1)){
+            this.array.index -= 1;
+        }else if ((index) < (this.getSize()-1)){
+            this.array.list[index] = this.array.list[index + 1];
+        } else {
+            throw new Exception("Index " + index + " out of bounds for size: " + this.getSize());
+        }
     }
 
     public void changePriority(int index, int new_Value) throws Exception {
-        int old_Value = this.HeapArray.list[index];
-        this.HeapArray.list[index] = new_Value;
-        if (new_Value > old_Value){
-            this.swiftUp(index);
-        } else {
-            this.swiftDown(index);
-        }
-    }
-
-    public void makeHeap() throws Exception {
-        int node_interno_index = this.parent(this.getSize()-1);
-        for (int i = node_interno_index; i >= 0; i--){
-            this.swiftDown(i);
-        }
-    }
-
-    public void heapSort() throws Exception {
-        int size = this.getSize();
-        for (int i = 0; i < size; i++){
-            int value = this.extractMax();
-            this.HeapArray.list[size-(i+1)] = value;
-        }
-        this.HeapArray.index = size;
-        for (int i = 0; i < size/2; i++){
-            int temp = this.HeapArray.list[i];
-            this.HeapArray.list[i] = this.HeapArray.list[size-(i+1)];
-            this.HeapArray.list[size-(i+1)] = temp;
-        }
-    }
-
-    public void heapSortRev() throws Exception {
-        int size = this.getSize();
-        for (int i = 0; i < size; i++){
-            int value = this.extractMax();
-            this.HeapArray.list[size-(i+1)] = value;
-        }
-        this.HeapArray.index = size;
-    }
-
-    public static void main(String[] args) throws Exception {
-        Heap myHeap = new Heap();
-        String data = "15 10 8 21 4";
-        String[] dataList = data.split(" ");
-        for (String s : dataList){
-            myHeap.HeapArray.pushBack(Integer.parseInt(s));
-        }
-        System.out.println(myHeap.levelOrder());
-        System.out.println("-------");
-        myHeap.makeHeap();
-        System.out.println(myHeap.levelOrder());
-        System.out.println("-------");
-        myHeap.heapSort();
-        System.out.println(myHeap.HeapArray);
-        System.out.println("-------");
-        myHeap.heapSortRev();
-        System.out.println(myHeap.HeapArray);
+        this.array.list[index] = new_Value;
     }
 }
